requirements.txt
cairo-lang==0.9.1
starknet.py
empiric-network
python-dotenv